pyproject.toml
[project]
name = "plex_sync"
version = "7.1.0"
description = "Sync watched between media servers locally"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "packaging==24.2",
    "plexapi==4.16.1",
    "pydantic==2.10.6",
    "python-dotenv==1.0.0",
    "requests==2.32.3",
    "croniter>=6.0.0",
]

[dependency-groups]
lint = [
    "ruff>=0.9.6",
]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.4",
    "types-requests>=2.32.0.20250306",
]
